#403 Split api into different modules
