+++ twake/backend/node/src/core/platform/framework/api/application-configuration.ts
@@ -1,0 +1,7 @@
+import { TwakeServiceOptions } from "./service-options";
+import { TwakeServiceConfiguration } from "./service-configuration";
+
+export class TwakeAppConfiguration extends TwakeServiceOptions<TwakeServiceConfiguration> {
+  services: Array<string>;
+  servicesPath: string;
+}
