+++ twake/backend/node/src/core/platform/framework/api/class.ts
@@ -1,0 +1,3 @@
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type Class = { new(...args: any[]): any; };
