+++ twake/backend/node/src/core/platform/framework/api/context.ts
@@ -1,0 +1,5 @@
+import { TwakeServiceProvider } from "./service-provider";
+
+export interface TwakeContext {
+  getProvider<T extends TwakeServiceProvider>(name: string): T;
+}
