+++ twake/backend/node/src/core/platform/framework/api/service-configuration.ts
@@ -1,0 +1,3 @@
+export interface TwakeServiceConfiguration {
+  get<T>(name: string, defaultValue?: T): T;
+}
