+++ twake/backend/node/src/core/platform/framework/api/service-definition.ts
@@ -1,0 +1,6 @@
+import { Class } from "./class";
+
+export interface ServiceDefinition {
+  name: string;
+  clazz: Class;
+}
