+++ twake/backend/node/src/core/platform/framework/api/service-interface.ts
@@ -1,0 +1,8 @@
+import { TwakeServiceProvider } from "./service-provider";
+
+export interface TwakeServiceInterface<T extends TwakeServiceProvider> {
+  doInit(): Promise<this>;
+  doStart(): Promise<this>;
+  doStop(): Promise<this>;
+  api(): T
+}
