+++ twake/backend/node/src/core/platform/framework/api/service-options.ts
@@ -1,0 +1,5 @@
+export class TwakeServiceOptions<TwakeServiceConfiguration> {
+  name?: string;
+  // TODO: configuration is abstract and comes from all others
+  configuration?: TwakeServiceConfiguration;
+}
