+++ twake/backend/node/src/core/platform/framework/api/service-provider.ts
@@ -1,0 +1,3 @@
+export interface TwakeServiceProvider {
+  version: string;
+}
