+++ twake/backend/node/src/core/platform/framework/api/service-state.ts
@@ -1,0 +1,10 @@
+export enum TwakeServiceState {
+  Ready = "READY",
+  Initializing = "INITIALIZING",
+  Initialized = "INITIALIZED",
+  Starting = "STARTING",
+  Started = "STARTED",
+  Stopping = "STOPPING",
+  Stopped = "STOPPED",
+  Errored = "ERRORED"
+}
